[security] Fix all apparent problematic alerts from CodeQL.

+++ common/to.etc.alg/src/main/java/to/etc/sjit/Lanczos3Filter.java
@@ -46,7 +46,7 @@ public class Lanczos3Filter extends ResamplerFilter {
 		if(value == 0.0f)
 			return 1.0f;
 
-		value *= Math.PI;
+		value *= (float) Math.PI;
 		return (float) Math.sin(value) / value;
 	}
